2.1 done, 2.2 halfway there
~ 2.1 ~
I made a function to count the number of characters in a name, using that function to count the number of characters of a name on the list, before comparing that against the highest and lowest character count. If it beats one of them, the number of characters becomes the new character count and the name is saved. Once the code has gone through all the names, it prints out the longest name and the shortest name as well as the total number of names.

~ 2.2 ~
I asked for the input of a name. I then used the character counter to count the number of characters in the name before subtracting to account for the fact that character count starts from 0 and not 1. A loop is then created where as long as the second character count is lower or the same as the total number of characters, it keeps going. I have 2 variables, with 'a' starting at 0 (the first letter) and 'b' starting at 2 (the second letter). A list that stores the pairs has been created and the variables are appended to the list as a pair. I then add 1 to each letter everytime the code loops so it slowly moves forward letter until the loop breaks. The resulting list of pairs is then printed. I still have the second part of 2.2 to do, I may skip it and leave it till later