Finishing 2.2
I added the last bit which was as simple as placing an append at the start and end of the loop, outside of it so it would add the first and last letter. I also forgot to mention last commit I had added a menu that I will add to as I build more functions. Currently there's a welcome message, followed by the menu. Exiting is simple with the user just needing to type 0 on the menu to quit