Completed 2.3
I have completed most of 2.3 by calculating the probabilities of different pairs of letters in a name giving me the whole name. I did this by creating 2 lists, one that contained how many pairs started with a certain letter, and another that contained how many times each pair occurred, with them being the denominator and numerator respectively. I would retrieve the numerator from the numerator list by using the pair as the key. I would then use the first character of the pair as the key to retrieve the denominator. Divide the numerator by the denominator and I get the probability. I then append the probability to a list that will be used to calculate the final probability. I print out the probability and the process starts again until the last letter. I then multiply all of the probabilities together to make the final probability, which I then print.